challenge-2/requirements.txt
# Medical Tissue Classification - Requirements
# Compatible with Google Colab, Kaggle, and local environments

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional: Additional models and augmentation
# timm>=0.9.0  # For more pretrained models
# albumentations>=1.3.0  # Advanced augmentation

# Notes:
# - Google Colab has most packages pre-installed
# - Kaggle notebooks also have these pre-installed
# - For local: pip install -r requirements.txt
